added action_get_test.go
